refactor: убрать зависимость InMemoryUsersStorage от FriendStorage
